popup windows done

+++ UI/functions.py
@@ -3,7 +3,7 @@ import os, sys
 import sqlite3
 from sqlite3 import Error
 #Pyside6 import
-from PySide6.QtWidgets import QTableWidgetItem
+from PySide6.QtWidgets import QTableWidgetItem, QMessageBox, QDialogButtonBox, QLabel,QVBoxLayout
 
 class AppFunctions():
     def __init__(self, arg):
@@ -91,7 +91,7 @@ class AppFunctions():
         phone = self.ui.phone.text()
         address = self.ui.address.text()
 
-        if nif != '' and userName !=  '' and email !=  '' and phone !=  '' and address !=  '':
+        if nif != '' and userName !=  '':
             add_user = f"""INSERT INTO Users (USER_NIF, USER_NAME, USER_EMAIL, USER_ADDRESS, USER_PHONE) VALUES('{nif}','{userName}','{email}','{phone}','{address}')"""
 
             if not conn.cursor().execute(add_user):
@@ -109,23 +109,80 @@ class AppFunctions():
                 AppFunctions.displayUsers(self, AppFunctions.getAllUsers(dbFolder))
         else:
             print("ERROR! Empty User Input")
+            dlg = QMessageBox(self)
+            dlg.setWindowTitle("Borrar usuario?")
+            dlg.setText("Por favor introduce un dato valido")
+            dlg.setStandardButtons(QMessageBox.Ok)
+            dlg.setIcon(QMessageBox.Warning)
 
+            dlg.exec()
+    
     def deleteUser(self, dbFolder):
         conn = AppFunctions.create_connection(dbFolder)
 
         nif = self.ui.nif_2.text()
 
         if nif != '':
-            delete_user = f"DELETE FROM Users WHERE USER_NIF='{nif}'"
+            
+            select_user=f"SELECT USER_NAME FROM Users WHERE USER_NIF='{nif}'"
 
-            if not conn.cursor().execute(delete_user):
-                print("ERROR! Could not delete user")
+            c = conn.cursor()
+            
+            # the user exists
+            if c.execute(select_user):
+
+                name = c.fetchall()[0][0]
+
+                dlg = QMessageBox(self)
+                dlg.setWindowTitle("Borrar usuario?")
+                dlg.setText(f"Seguro que quieres borrar a {name}??")
+                dlg.setStandardButtons(QMessageBox.Yes | QMessageBox.No)
+                dlg.setIcon(QMessageBox.Question)
+
+                btn = dlg.exec()
+
+                if btn == QMessageBox.Yes:
+                    delete_user = f"DELETE FROM Users WHERE USER_NIF='{nif}'"
+
+                    if not conn.cursor().execute(delete_user):
+                        print("ERROR! Could not delete user")
+                        dlg.setText("No se ha podido borrar el usuario")
+                        dlg.setStandardButtons(QMessageBox.Ok)
+                        dlg.setIcon(QMessageBox.Warning)
+                    else:
+                        print("DELETE USER OK!")
+                        dlg2 = QMessageBox(self)
+                        dlg2.setText("El usuario se ha borrado correctamente, se actualizará cuando reinicies la aplicación")
+                        dlg2.setStandardButtons(QMessageBox.Ok)
+                        dlg2.setIcon(QMessageBox.Information)
+                        dlg2.exec()
+                        conn.commit()
+                        #clear from input
+                        self.ui.nif_2.setText("")
+
+                else:
+                    print(f"CONTIUNE! The user {name} will not be deleted")
             else:
-                print("DELETE USER OK!")
-                conn.commit()
-                #clear from input
-                self.ui.nif_2.setText("")
+                print("ERROR! The user does not exist!")
+                dlg = QMessageBox(self)
+                dlg.setWindowTitle("Borrar usuario?")
+                dlg.setText("El usuario introducido no existe en la base de datos")
+                dlg.setStandardButtons(QMessageBox.Ok)
+                dlg.setIcon(QMessageBox.Warning)
 
-                AppFunctions.displayUsers(self, AppFunctions.getAllUsers(dbFolder))
+                btn = dlg.exec()
         else:
-            print("ERROR! Empty User Input")
+            print("Empty Input")
+            dlg = QMessageBox(self)
+            dlg.setWindowTitle("Borrar usuario?")
+            dlg.setText("Por favor introduce un dato valido")
+            dlg.setStandardButtons(QMessageBox.Ok)
+            dlg.setIcon(QMessageBox.Warning)
+
+            dlg.exec()
+        
+        #Always show the users 
+        AppFunctions.displayUsers(self, AppFunctions.getAllUsers(dbFolder))
+
+    def updateUser(self, dbFolder):
+        print("Update")
